organized character form files into subfolders of NewCharacterForm folder.
