+++ src/components/newCharacterForm/NewCharacterFormContainer.js
@@ -1,10 +1,10 @@
 import { useState, useEffect } from "react"
-import { SpeciesSelection } from "./SpeciesSelection"
-import { BackgroundSelection } from "./BackgroundSelection"
-import { ClassSelection } from "./ClassSelection"
-import { AttributeSelection } from "./AttributeSelection"
-import { EquipmentSelection } from "./EquipmentSelection"
-import { CharacterInfoSelection } from "./CharacterInfoSelection"
+import { SpeciesSelection } from "./species/SpeciesSelection"
+import { BackgroundSelection } from "./background/BackgroundSelection"
+import { ClassSelection } from "./class/ClassSelection"
+import { AttributeSelection } from "./attributes/AttributeSelection"
+import { EquipmentSelection } from "./equipment/EquipmentSelection"
+import { CharacterInfoSelection } from "./information/CharacterInfoSelection"
 import { CharacterSubmitButton } from "./CharacterSubmitButton"
 import { getAllAttributesFetch } from "../ApiManager"
 import { StepButtons } from "./StepButtons"
