+++ src/components/newCharacterForm/attributes/AttributeValueSelection.js
